Level Select Map
The player is now able to select which levels to play through the level select map. The player is able to select the level by going to the level  (WASD) and selecting the map with the jump button (Space Bar). If the level is locked the player is unable to play the level.