Switching Music and Continue Button in Title Screen
Music now switches to level complete music when the player completes the level. The continue button and new game button is added to the title screen. The new game button erases the player prefs data to create a new game. the continue button only appears if the player has a saved game data.